requirements.txt
numpy
pandas
pyyaml
anthropic
openai
google-genai
python-dotenv
tqdm
